feat: RobokIDE modularization
