+++ feature/feature-setting/src/main/java/org/gampiot/robokide/feature/settings/ui/fragment/SettingsFragment.kt
@@ -11,7 +11,7 @@ import org.gampiot.robokide.feature.settings.R
 import org.gampiot.robokide.feature.settings.databinding.FragmentSettingsBinding
 import org.gampiot.robokide.feature.base.ui.RobokFragment
 import org.gampiot.robokide.feature.settings.ui.fragment.editor.SettingsEditorFragment
-import org.gampiot.robokide.features.res.strings.Strings
+import org.gampiot.robokide.feature.res.strings.Strings
 
 import dev.trindadedev.lib.ui.components.preferences.withicon.Preference
