+++ robok/robok-compiler/build.gradle.kts
@@ -44,8 +44,8 @@ tasks.withType<org.jetbrains.kotlin.gradle.tasks.KotlinCompile>().configureEach 
 }
 
 dependencies {
-    implementation(project(":robok:util"))
-    implementation(project(":robok:lang"))
+    implementation(project(":robok:robok-util"))
+    implementation(project(":robok:robok-lang"))
 }
 
 publishing {
