+++ robok/robok-compiler/src/main/AndroidManifest.xml
