+++ robok/robok-diagnostic/src/main/AndroidManifest.xml
