+++ robok/robok-diagnostic/src/main/java/robok/diagnostic/logic/DiagnosticHandler.kt
