+++ robok/robok-diagnostic/src/main/java/robok/diagnostic/logic/Java8BaseVisitor.java
