+++ robok/robok-util/src/main/AndroidManifest.xml
